refactor(beaconevent): extract classification into classify.go
Final converter-split PR. Move the event-classification code into a new
same-package classify.go: the action vocabulary const block, harnessNameRules
and textKeywordActionRules tables, HarnessName, NormalizeHarnessName, InferAction,
HasToolCall, HasPromptLikeContent, CopilotAction, GeminiToolAction,
GeminiFileAction, EventCategory, and the IsMeaningfulValue/IsFileURI/
FilePathFromURI helpers used by them.

converter.go now holds the core conversion (Converter, EventsFrom*/EventFrom*,
ShouldDrop*, codex/vscode normalization, metric usage, PopulateCommon,
MCPFromAttrs, populateRunContext, RawPayload) at ~711 lines, down from 1613.
Move-only, behavior-preserving: tests stay green.
